% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparingstatuses.R
\name{comparingstatuses}
\alias{comparingstatuses}
\title{Comparingstatuses}
\usage{
comparingstatuses(
  data,
  state1 = NULL,
  state2 = NULL,
  state3 = NULL,
  state4 = NULL,
  comparison = NULL
)
}
\arguments{
\item{data}{data.frame. Needs to be the output from generatinginfosheet, with statuses.}

\item{state1}{character string. The state the polygon is in.}

\item{state2}{character string. If the polygon is in 2 states, the second state the polygon is in.}

\item{state3}{character string. If the polygon is in 3 states, the third state the polygon is in.}

\item{state4}{character string. If the polygon is in 4 states, the fourth state the polygon is in. The limit is 4 states.}

\item{comparison}{character string. Either "most concern" or "least concern". If "most concern", the final_status will be the status that is of the most concern. If "least concern", the final_status will be the status of least concern.}
}
\value{
A dataframe with a Final status for each species.
}
\description{
This function compares the species' statuses from different organizations
to determine the finsl status to be used for each species.
}
