% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_biodiversitydata.R
\name{filter_biodiversitydata}
\alias{filter_biodiversitydata}
\title{filter_biodiversitydata}
\usage{
filter_biodiversitydata(
  data,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  species = "species",
  countries = NULL,
  tests = c("capitals", "centroids", "equal", "gbif", "institutions", "outliers", "seas",
    "zeros"),
  capitals_rad = 10000,
  centroids_rad = 1000,
  centroids_detail = "both",
  inst_rad = 100,
  outliers_method = "quantile",
  outliers_mtp = 5,
  outliers_td = 1000,
  outliers_size = 7,
  range_rad = 0,
  zeros_rad = 0.5,
  capitals_ref = NULL,
  centroids_ref = NULL,
  country_ref = NULL,
  country_refcol = "iso_a3",
  country_buffer = NULL,
  inst_ref = NULL,
  range_ref = NULL,
  seas_ref = NULL,
  seas_scale = 50,
  seas_buffer = NULL,
  urban_ref = NULL,
  aohi_rad = NULL,
  value = "spatialvalid",
  verbose = TRUE,
  report = FALSE,
  coordinate_uncertainty_threshold = 1610,
  human_observation = TRUE,
  observation = TRUE,
  preserved_specimen = TRUE,
  machine_observation = TRUE,
  material_sample = TRUE,
  living_specimen = TRUE,
  occurrence = TRUE,
  na_treatment_precision = TRUE,
  na_treatment_base = TRUE,
  output_incertae_file = "incertaesedis.csv",
  output_clean_file = "cleanrecords.csv",
  output_flagged_file = "flaggedrecords.csv"
)
}
\arguments{
\item{data}{data.frame. Containing geographical coordinates and species names.}

\item{lon}{character string. The column with the longitude coordinates. Default = “decimalLongitude”.}

\item{lat}{character string. The column with the latitude coordinates. Default = “decimalLatitude”.}

\item{species}{a character string. A vector of the same length as rows in x, with the species identity for each record. If NULL, tests must not include the "outliers" or "duplicates" tests.}

\item{countries}{a character string. The column with the country assignment of each record in three letter ISO code. Default = “countrycode”. If missing, the countries test is skipped.}

\item{tests}{tests    a vector of character strings, indicating which tests to run. See details for all tests available. Default = c("capitals", "centroids", "equal", "gbif", "institutions", "outliers", "seas", "zeros")}

\item{centroids_rad}{numeric. The radius around centroid coordinates in meters. Default = 1000.}

\item{centroids_detail}{a character string. If set to ‘country’ only country (adm-0) centroids are tested, if set to ‘provinces’ only province (adm-1) centroids are tested. Default = ‘both’.}

\item{inst_rad}{numeric. The radius around biodiversity institutions coordinates in metres. Default = 100.}

\item{outliers_method}{The method used for outlier testing. See details.}

\item{outliers_mtp}{numeric. The multiplier for the interquartile range of the outlier test. If NULL outliers.td is used. Default = 5.}

\item{outliers_td}{numeric. The minimum distance of a record to all other records of a species to be identified as outlier, in km. Default = 1000.}

\item{outliers_size}{numeric. The minimum number of records in a dataset to run the taxon-specific outlier test. Default = 7.}

\item{range_rad}{buffer around natural ranges. Default = 0.}

\item{zeros_rad}{numeric. The radius around 0/0 in degrees. Default = 0.5.}

\item{centroids_ref}{a data.frame with alternative reference data for the centroid test. If NULL, the countryref dataset is used. Alternatives must be identical in structure.}

\item{country_ref}{a SpatVector as alternative reference for the countries test. If NULL, the rnaturalearth:ne_countries('medium', returnclass = "sf") dataset is used.}

\item{country_refcol}{the column name in the reference dataset, containing the relevant ISO codes for matching. Default is to "iso_a3_eh" which referes to the ISO-3 codes in the reference dataset. See notes.}

\item{country_buffer}{numeric. Units are in meters. If provided, a buffer is created around each country polygon.}

\item{inst_ref}{data.frame with alternative reference data for the biodiversity institution test. If NULL, the institutions dataset is used. Alternatives must be identical in structure.}

\item{range_ref}{a SpatVector of species natural ranges. Required to include the 'ranges' test. See cc_iucn for details.}

\item{seas_ref}{a SpatVector as alternative reference for the seas test. If NULL, the rnaturalearth::ne_download(scale = 110, type = 'land', category = 'physical', returnclass = "sf") dataset is used.}

\item{seas_scale}{The scale of the default landmass reference. Must be one of 10, 50, 110. Higher numbers equal higher detail. Default = 50.}

\item{seas_buffer}{numeric. Units are in meters. If provided, a buffer is created around sea polygon.}

\item{urban_ref}{a SpatVector as alternative reference for the urban test. If NULL, the test is skipped. See details for a reference gazetteers.}

\item{aohi_rad}{numeric. The radius around aohi coordinates in meters. Default = 1000.}

\item{value}{a character string defining the output value. See the value section for details. one of ‘spatialvalid’, ‘summary’, ‘clean’. Default = ‘spatialvalid’.}

\item{verbose}{logical. If TRUE reports the name of the test and the number of records flagged.}

\item{report}{logical or character. If TRUE a report file is written to the working directory, summarizing the cleaning results. If a character, the path to which the file should be written. Default = FALSE.}

\item{coordinate_uncertainty_threshold}{numeric. Threshold for coordinate uncertainty to be filtered out, in meters. Default = 1600 meters (1 mile).}

\item{human_observation}{logical. If FALSE human observations will be filtered out. Default = TRUE.}

\item{observation}{logical. If FALSE  observations will be filtered out. Default = TRUE.}

\item{preserved_specimen}{logical. If FALSE preserved specimen will be filtered out. Default = TRUE.}

\item{machine_observation}{logical. If FALSE machine observations will be filtered out. Default = TRUE.}

\item{material_sample}{logical. If FALSE material samples will be filtered out. Default = TRUE.}

\item{living_specimen}{logical. If FALSE living specimen will be filtered out. Default = TRUE.}

\item{occurrence}{logical. If FALSE occurrences will be filtered out. Default = TRUE.}

\item{na_treatment_precision}{logical. If FALSE NAs for coordinate precision will be filtered out. Default = TRUE.}

\item{na_treatment_base}{logical. If FALSE NAs for basis of record will be filtered out. Default = TRUE.}

\item{output_incertae_file}{character string. Name of the file to output Incertae sedis records to.}

\item{output_clean_file}{character string. Name of the file to output the clean records to.}

\item{output_flagged_file}{character string. Name of the file to output the flagged records to.}

\item{captials_rad}{numeric. The radius around capital coordinates in meters. Default = 10000.}

\item{captials_ref}{a data.frame with alternative reference data for the country capitals test. If missing, the countryref dataset is used. Alternatives must be identical in structure.}
}
\value{
A dataframe with a filtered biodiversity data.
}
\description{
This function filters GBIF data.
}
