% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatinginfosheet.R
\name{generatinginfosheet}
\alias{generatinginfosheet}
\title{GeneratingInfoSheet}
\usage{
generatinginfosheet(
  data,
  mark_invasive_species = TRUE,
  include_USinvasives_list = TRUE,
  state1 = NULL,
  state2 = NULL,
  state3 = NULL,
  state4 = NULL,
  natureserve_statuses = TRUE,
  ICUNRedList_statuses = TRUE
)
}
\arguments{
\item{data}{data.frame. Should be the filtered_data from filter_biodiversity_data.}

\item{mark_invasive_species}{logical. When true, invasive species will be identified and marked}

\item{include_USinvasives_list}{logical. When true, the USInvasive list will be included when identifying/marking invasive species. The US Invasive List is simply a list of the most invasive species in the US. Not state-specific.}

\item{state1}{character string. The state the polygon is in. Only need states if mark_invasive_species, natureserve_statuses, or ICUNRedList_statuses are marked TRUE.}

\item{state2}{character string. If the polygon is in 2 states, the second state the polygon is in.}

\item{state3}{character string.  If the polygon is in 3 states, the third state the polygon is in.}

\item{state4}{character string. If the piolygon is in 4 states, the fourth state the polygon is in. The limit is 4 states.}

\item{natureserve_statuses}{logical. When true, Natureserve statuses will be added to each species, for each state as well as the global status.}

\item{ICUNRedList_statuses}{logical. When true, ICUN Red List statuses will be added to each species.}
}
\value{
A dataframe with information for each species.
}
\description{
This function generates an information sheet for all of the species in the polygon.
}
